+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Targets
 Unreleased
 ----------
 * AC branch flow limits in terms of current magnitudes squared.
+* Support for init values for constraint extra variables (done).
+* Improved post processing of structures of constraint Hessians (now constr.c ensures lower triangular and fills H_combined ij, fixing bug in ac_flow_lim) (done).
 
 Version 1.3.0
 -------------